backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
psycopg2-binary==2.9.10
bcrypt>=4.0.0
python-jose[cryptography]==3.3.0
email-validator>=2.0.0
